+++ tp-api/src/main/java/com/thirdplace/ThirdPlaceDatabaseService/ResultType.java
@@ -1,4 +1,4 @@
-package com.thirdplace.ThirdPlaceDatabaseService;
+package com.thirdplace.thirdplacedatabaseservice;
 
 public interface ResultType {
