+++ tp-api/src/main/java/com/thirdplace/ThirdPlaceDatabaseService/TableColumnType.java
@@ -1,4 +1,4 @@
-package com.thirdplace.ThirdPlaceDatabaseService;
+package com.thirdplace.thirdplacedatabaseservice;
 
 /**
  * Class to represent a column type in a table
