+++ tp-api/src/test/java/com/thirdplace/usertabledriver/UserTableDriverTests.java
@@ -1,0 +1,5 @@
+package com.thirdplace.usertabledriver;
+
+public class UserTableDriverTests {
+    
+}
